+++ specs/example/StringUtils.spec.ts
@@ -11,8 +11,14 @@ describe('StringUtils', () => {
         });
 
         it('should throw an error if input is not a string', () => {
-            expect(() => StringUtils.reverse(123 as any)).to.throw(MError, 'Input must be a string');
-            expect(() => StringUtils.reverse(null as any)).to.throw(MError, 'Input must be a string');
+            expect(() => StringUtils.reverse(123 as any)).to.throw(
+                MError,
+                'Input must be a string'
+            );
+            expect(() => StringUtils.reverse(null as any)).to.throw(
+                MError,
+                'Input must be a string'
+            );
         });
     });
 
@@ -29,8 +35,14 @@ describe('StringUtils', () => {
         });
 
         it('should throw an error if input is not a string', () => {
-            expect(() => StringUtils.isPalindrome(123 as any)).to.throw(MError, 'Input must be a string');
-            expect(() => StringUtils.isPalindrome(null as any)).to.throw(MError, 'Input must be a string');
+            expect(() => StringUtils.isPalindrome(123 as any)).to.throw(
+                MError,
+                'Input must be a string'
+            );
+            expect(() => StringUtils.isPalindrome(null as any)).to.throw(
+                MError,
+                'Input must be a string'
+            );
         });
     });
 });